TASK 2_3_1 SECOND FIX

+++ src/main/java/PP231/dao/UserDao.java
@@ -8,9 +8,9 @@ public interface UserDao {
 
     void addUser(User user);
 
-    void deleteUser(Long id);
+    void deleteUser(int id);
 
     List<User> getAllUsers();
 
-    User getUserById(Long id);
+    User getUserById(int id);
 }
